+++ src/auth/auth.module.ts
@@ -5,9 +5,21 @@ import { AuthService } from './auth.service';
 import { JwtStrategy } from './strategy/jwt.strategy';
 
 import { UserModule } from '../feature/user/user.app.module';
+import { JwtModule } from '@nestjs/jwt';
+import { AuthController } from './auth.controller';
+import { ConfigModule } from '@nestjs/config';
 
 @Module({
-  imports: [PassportModule, UserModule],
+  imports: [
+    ConfigModule.forRoot(),
+    UserModule,
+    PassportModule,
+    JwtModule.register({
+      secret: process.env.JWT_SECRET,
+      signOptions: { expiresIn: '24h' }, // 24시간
+    }),
+  ],
+  controllers: [AuthController],
   providers: [AuthService, JwtStrategy],
 })
 export class AuthModule {}
