+++ src/feature/user/user.app.module.ts
@@ -1,7 +1,5 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { JwtModule } from '@nestjs/jwt';
-import { ConfigModule } from '@nestjs/config';
 
 import { User } from '../../entity/user.entity';
 import { UserRepository } from './repository/user.repository';
@@ -10,14 +8,7 @@ import { UserService } from './user.service';
 import { UserLib } from './user.lib';
 
 @Module({
-  imports: [
-    ConfigModule.forRoot(), // TODO: configuration 파일 형식으로 정리 필요해보임
-    TypeOrmModule.forFeature([User]),
-    JwtModule.register({
-      secret: process.env.JWT_SECRET,
-      signOptions: { expiresIn: '24h' }, // 24시간
-    }),
-  ],
+  imports: [TypeOrmModule.forFeature([User])],
   providers: [UserService, UserRepository, UserLib],
   controllers: [UserController],
   exports: [UserLib],
